+++ src/manifold/src/quickhull.h
@@ -55,18 +55,10 @@
  * */
 #pragma once
 #include <array>
-#include <cassert>
-#include <cinttypes>
-#include <cmath>
 #include <deque>
-#include <fstream>
-#include <iostream>
 #include <limits>
-#include <memory>
-#include <unordered_map>
 #include <vector>
 
-#include "par.h"
 #include "shared.h"
 #include "vec.h"
 
@@ -82,7 +74,6 @@ class ConvexHull {
 };
 
 // Pool.hpp
-
 class Pool {
   std::vector<std::unique_ptr<Vec<size_t>>> data;
 
@@ -126,9 +117,7 @@ class Plane {
 
   // Construct a plane using normal N and any point P on the plane
   Plane(const glm::dvec3& N, const glm::dvec3& P)
-      : N(N),
-        D(glm::dot(-N, P)),
-        sqrNLength(N.x * N.x + N.y * N.y + N.z * N.z) {}
+      : N(N), D(glm::dot(-N, P)), sqrNLength(glm::dot(N, N)) {}
 };
 
 // Ray.hpp
@@ -166,21 +155,24 @@ class MeshBuilder {
   struct Face {
     size_t he;
     Plane P{};
-    double mostDistantPointDist;
-    size_t mostDistantPoint;
-    size_t visibilityCheckedOnIteration;
+    double mostDistantPointDist = 0.0;
+    size_t mostDistantPoint = 0;
+    size_t visibilityCheckedOnIteration = 0;
     std::uint8_t isVisibleFaceOnCurrentIteration : 1;
     std::uint8_t inFaceStack : 1;
-    std::uint8_t horizonEdgesOnCurrentIteration : 3;
     // Bit for each half edge assigned to this face, each being 0 or 1 depending
     // on whether the edge belongs to horizon edge
+    std::uint8_t horizonEdgesOnCurrentIteration : 3;
     std::unique_ptr<Vec<size_t>> pointsOnPositiveSide;
+
+    Face(size_t he)
+        : he(he),
+          isVisibleFaceOnCurrentIteration(0),
+          inFaceStack(0),
+          horizonEdgesOnCurrentIteration(0) {}
 
     Face()
         : he(std::numeric_limits<size_t>::max()),
-          mostDistantPointDist(0),
-          mostDistantPoint(0),
-          visibilityCheckedOnIteration(0),
           isVisibleFaceOnCurrentIteration(0),
           inFaceStack(0),
           horizonEdgesOnCurrentIteration(0) {}
@@ -201,29 +193,9 @@ class MeshBuilder {
   // the following vectors.
   std::vector<size_t> disabledFaces, disabledHalfEdges;
 
-  size_t addFace() {
-    if (disabledFaces.size()) {
-      size_t index = disabledFaces.back();
-      auto& f = faces[index];
-      assert(f.isDisabled());
-      assert(!f.pointsOnPositiveSide);
-      f.mostDistantPointDist = 0;
-      disabledFaces.pop_back();
-      return index;
-    }
-    faces.emplace_back();
-    return faces.size() - 1;
-  }
+  size_t addFace();
 
-  size_t addHalfEdge() {
-    if (disabledHalfEdges.size()) {
-      const size_t index = disabledHalfEdges.back();
-      disabledHalfEdges.pop_back();
-      return index;
-    }
-    halfEdges.emplace_back();
-    return halfEdges.size() - 1;
-  }
+  size_t addHalfEdge();
 
   // Mark a face as disabled and return a pointer to the points that were on the
   // positive of it.
@@ -244,69 +216,9 @@ class MeshBuilder {
 
   // Create a mesh with initial tetrahedron ABCD. Dot product of AB with the
   // normal of triangle ABC should be negative.
-  void setup(size_t a, size_t b, size_t c, size_t d) {
-    faces.clear();
-    halfEdges.clear();
-    disabledFaces.clear();
-    disabledHalfEdges.clear();
+  void setup(size_t a, size_t b, size_t c, size_t d);
 
-    faces.reserve(4);
-    halfEdges.reserve(12);
-
-    // Create halfedges
-    // AB
-    halfEdges.push_back(HalfEdge(b, 6, 0, 1));
-    // BC
-    halfEdges.push_back(HalfEdge(c, 9, 0, 2));
-    // CA
-    halfEdges.push_back(HalfEdge(a, 3, 0, 0));
-    // AC
-    halfEdges.push_back(HalfEdge(c, 2, 1, 4));
-    // CD
-    halfEdges.push_back(HalfEdge(d, 11, 1, 5));
-    // DA
-    halfEdges.push_back(HalfEdge(a, 7, 1, 3));
-    // BA
-    halfEdges.push_back(HalfEdge(a, 0, 2, 7));
-    // AD
-    halfEdges.push_back(HalfEdge(d, 5, 2, 8));
-    // DB
-    halfEdges.push_back(HalfEdge(b, 10, 2, 6));
-    // CB
-    halfEdges.push_back(HalfEdge(b, 1, 3, 10));
-    // BD
-    halfEdges.push_back(HalfEdge(d, 8, 3, 11));
-    // DC
-    halfEdges.push_back(HalfEdge(c, 4, 3, 9));
-
-    // Create faces
-    Face ABC;
-    ABC.he = 0;
-    faces.push_back(std::move(ABC));
-
-    Face ACD;
-    ACD.he = 3;
-    faces.push_back(std::move(ACD));
-
-    Face BAD;
-    BAD.he = 6;
-    faces.push_back(std::move(BAD));
-
-    Face CBD;
-    CBD.he = 9;
-    faces.push_back(std::move(CBD));
-  }
-
-  std::array<size_t, 3> getVertexIndicesOfFace(const Face& f) const {
-    std::array<size_t, 3> v;
-    const HalfEdge* he = &halfEdges[f.he];
-    v[0] = he->endVertex;
-    he = &halfEdges[he->next];
-    v[1] = he->endVertex;
-    he = &halfEdges[he->next];
-    v[2] = he->endVertex;
-    return v;
-  }
+  std::array<size_t, 3> getVertexIndicesOfFace(const Face& f) const;
 
   std::array<size_t, 2> getVertexIndicesOfHalfEdge(const HalfEdge& he) const {
     return {halfEdges[he.opp].endVertex, he.endVertex};
@@ -334,6 +246,8 @@ class HalfEdgeMesh {
   struct Face {
     // Index of one of the half edges of this face
     size_t halfEdgeIndex;
+
+    Face(size_t halfEdgeIndex) : halfEdgeIndex(halfEdgeIndex) {}
   };
 
   std::vector<glm::dvec3> vertices;
@@ -341,94 +255,13 @@ class HalfEdgeMesh {
   std::vector<HalfEdge> halfEdges;
 
   HalfEdgeMesh(const MeshBuilder& builderObject,
-               const VecView<glm::dvec3>& vertexData) {
-    std::unordered_map<size_t, size_t> faceMapping;
-    std::unordered_map<size_t, size_t> halfEdgeMapping;
-    std::unordered_map<size_t, size_t> vertexMapping;
-
-    size_t i = 0;
-    for (const auto& face : builderObject.faces) {
-      if (!face.isDisabled()) {
-        faces.push_back({static_cast<size_t>(face.he)});
-        faceMapping[i] = faces.size() - 1;
-
-        const auto heIndices = builderObject.getHalfEdgeIndicesOfFace(face);
-        for (const auto heIndex : heIndices) {
-          const size_t vertexIndex = builderObject.halfEdges[heIndex].endVertex;
-          if (vertexMapping.count(vertexIndex) == 0) {
-            vertices.push_back(vertexData[vertexIndex]);
-            vertexMapping[vertexIndex] = vertices.size() - 1;
-          }
-        }
-      }
-      i++;
-    }
-
-    i = 0;
-    for (const auto& halfEdge : builderObject.halfEdges) {
-      if (!halfEdge.isDisabled()) {
-        halfEdges.push_back(HalfEdge(halfEdge.endVertex, halfEdge.opp,
-                                     halfEdge.face, halfEdge.next));
-        halfEdgeMapping[i] = halfEdges.size() - 1;
-      }
-      i++;
-    }
-
-    for (auto& face : faces) {
-      assert(halfEdgeMapping.count(face.halfEdgeIndex) == 1);
-      face.halfEdgeIndex = halfEdgeMapping[face.halfEdgeIndex];
-    }
-
-    for (auto& he : halfEdges) {
-      he.face = faceMapping[he.face];
-      he.opp = halfEdgeMapping[he.opp];
-      he.next = halfEdgeMapping[he.next];
-      he.endVertex = vertexMapping[he.endVertex];
-    }
-  }
+               const VecView<glm::dvec3>& vertexData);
 };
-
-// MathUtils.hpp
-
-namespace mathutils {
-
-inline double getSquaredDistanceBetweenPointAndRay(const glm::dvec3& p,
-                                                   const Ray& r) {
-  const glm::dvec3 s = p - r.S;
-  double t = glm::dot(s, r.V);
-  return glm::dot(s, s) - t * t * r.VInvLengthSquared;
-}
-
-inline double getSquaredDistance(const glm::dvec3& p1, const glm::dvec3& p2) {
-  return glm::dot(p1 - p2, p1 - p2);
-}
-// Note that the unit of distance returned is relative to plane's normal's
-// length (divide by N.getNormalized() if needed to get the "real" distance).
-inline double getSignedDistanceToPlane(const glm::dvec3& v, const Plane& p) {
-  return glm::dot(p.N, v) + p.D;
-}
-
-inline glm::dvec3 getTriangleNormal(const glm::dvec3& a, const glm::dvec3& b,
-                                    const glm::dvec3& c) {
-  // We want to get (a-c).crossProduct(b-c) without constructing temp vectors
-  double x = a.x - c.x;
-  double y = a.y - c.y;
-  double z = a.z - c.z;
-  double rhsx = b.x - c.x;
-  double rhsy = b.y - c.y;
-  double rhsz = b.z - c.z;
-  double px = y * rhsz - z * rhsy;
-  double py = z * rhsx - x * rhsz;
-  double pz = x * rhsy - y * rhsx;
-  return glm::normalize(glm::dvec3(px, py, pz));
-}
-
-}  // namespace mathutils
 
 // QuickHull.hpp
 
 struct DiagnosticsData {
-  // How many times QuickHull failed to solve the  horizon edge. Failures lead
+  // How many times QuickHull failed to solve the horizon edge. Failures lead
   // to degenerated convex hulls.
   size_t failedHorizonEdges;
 
@@ -527,45 +360,4 @@ class QuickHull {
   // Get diagnostics about last generated convex hull
   const DiagnosticsData& getDiagnostics() { return diagnostics; }
 };
-
-/*
- * Inline function definitions
- */
-
-std::unique_ptr<Vec<size_t>> QuickHull::getIndexVectorFromPool() {
-  auto r = indexVectorPool.get();
-  // r->clear();
-  r->resize(0);
-  return r;
-}
-
-void QuickHull::reclaimToIndexVectorPool(std::unique_ptr<Vec<size_t>>& ptr) {
-  const size_t oldSize = ptr->size();
-  if ((oldSize + 1) * 128 < ptr->capacity()) {
-    // Reduce memory usage! Huge vectors are needed at the beginning of
-    // iteration when faces have many points on their positive side. Later on,
-    // smaller vectors will suffice.
-    ptr.reset(nullptr);
-    return;
-  }
-  indexVectorPool.reclaim(ptr);
-}
-
-bool QuickHull::addPointToFace(typename MeshBuilder::Face& f,
-                               size_t pointIndex) {
-  const double D =
-      mathutils::getSignedDistanceToPlane(originalVertexData[pointIndex], f.P);
-  if (D > 0 && D * D > epsilonSquared * f.P.sqrNLength) {
-    if (!f.pointsOnPositiveSide) {
-      f.pointsOnPositiveSide = getIndexVectorFromPool();
-    }
-    f.pointsOnPositiveSide->push_back(pointIndex);
-    if (D > f.mostDistantPointDist) {
-      f.mostDistantPointDist = D;
-      f.mostDistantPoint = pointIndex;
-    }
-    return true;
-  }
-  return false;
-}
 }  // namespace manifold
